authentication

+++ resources/item.py
@@ -1,7 +1,7 @@
-import uuid
 from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
+from flask_jwt_extended import jwt_required, get_jwt
 # Todos los errores de SQLAlchemy heredan de este error
 from sqlalchemy.exc import SQLAlchemyError
 
@@ -12,16 +12,21 @@ from schemas import ItemSchema, ItemUpdateSchema
 blp = Blueprint("Items", __name__, description="Operations on items")
 
 
-@blp.route("/item/<string:item_id>")
+@blp.route("/item/<int:item_id>")
 class Item(MethodView):
     # Decorate main success response, pasa lo que retornemos por ItemSchema,
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):  # Endpoint to get an item
         # Query available with Flask-SQLAlchemy and not with VAnilla SQLAlchemy
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required()
     def delete(self, item_id):  # Endpoint to delete an item
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -46,10 +51,12 @@ class Item(MethodView):
 
 @blp.route("/item")
 class ItemList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):  # Endpoint to get all items
         return ItemModel.query.all()
 
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):  # Endpoint to create an item
